Arreglo errores debidos a Fecha

+++ src/main/java/co/edu/uniandes/dse/museoartemoderno/services/ArtistaService.java
@@ -1,5 +1,6 @@
 package co.edu.uniandes.dse.museoartemoderno.services;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,6 +10,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import co.edu.uniandes.dse.museoartemoderno.repositories.ArtistaRepository;
 import co.edu.uniandes.dse.museoartemoderno.entities.ArtistaEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.MovimientoArtisticoEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.MuseoEntity;
+import co.edu.uniandes.dse.museoartemoderno.entities.ObraEntity;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.EntityNotFoundException;
 import co.edu.uniandes.dse.museoartemoderno.exceptions.IllegalOperationException;
 
@@ -32,7 +36,38 @@ public class ArtistaService {
 	public ArtistaEntity createArtista(ArtistaEntity artistaEntity) throws EntityNotFoundException, IllegalOperationException {
 		log.info("Inicia proceso de creación del artista");
 		
-			
+		if (!validateNombre(artistaEntity.getNombre()))
+			throw new IllegalOperationException("Nombre is not valid");	
+		
+		if (!artistaRepository.findByNombre(artistaEntity.getNombre()).isEmpty())
+			throw new IllegalOperationException("Nombre already exists");
+		
+		if (artistaEntity.getFechaNacimiento() == null)
+			throw new IllegalOperationException("Fecha Nacimiento is not valid");
+		
+		if (!validateFechaNacimiento(artistaEntity.getFechaNacimiento()))
+			throw new IllegalOperationException("Fecha Nacimiento is not valid");
+		
+		if (artistaEntity.getFechaFallecimiento() == null)
+			throw new IllegalOperationException("Fecha Fallecimiento is not valid");
+		
+		if (!validateFechaFallecimiento(artistaEntity.getFechaFallecimiento()))
+			throw new IllegalOperationException("Fecha Fallecimiento is not valid");	
+		
+		if (artistaEntity.getLugarNacimiento() == null)
+			throw new IllegalOperationException("Lugar Nacimiento is not valid");	
+		
+		if (artistaEntity.getLugarFallecimiento() == null)
+			throw new IllegalOperationException("Lugar Fallecimiento is not valid");
+		
+		if (artistaEntity.getMovimientos() == null)
+			throw new IllegalOperationException("Movimientos is not valid");
+		
+		if (artistaEntity.getMuseos() == null)
+			throw new IllegalOperationException("Museos is not valid");
+		
+		if (artistaEntity.getObras() == null)
+			throw new IllegalOperationException("Obras is not valid");
 		
 		log.info("Termina proceso de creación del artista");
 		return artistaRepository.save(artistaEntity);
@@ -65,27 +100,52 @@ public class ArtistaService {
 		log.info("Termina proceso de consultar el artista con id: " + artistaId);
 		return artistaEntity.get();
 	}
-	
 	/**
-	 * Actualizar artista dado su Id
-	 * @param artistaId - Id del artista que se quiere actualizar
-	 * @param artista - entidad con los cambios
-	 * @return - entidad actualizada
+	 * Actualiza los datos de un artista
+	 * @param artistaId - id del artista que se quiere actualizar
+	 * @param artista - la entidad artista con los datos nuevos
+	 * @return La entidad del artista actualizada
 	 * @throws EntityNotFoundException - Exception que se lanza si no se encuentra la entidad
 	 * @throws IllegalOperationException - Exception que se lanza si no se cumple alguna regla de negocio
 	 */
 	@Transactional
 	public ArtistaEntity updateArtista(Long artistaId, ArtistaEntity artista)
 			throws EntityNotFoundException, IllegalOperationException {
-		log.info("Inicia proceso de actualizar el libro con id: " + artistaId);
+		log.info("Inicia proceso de actualizar el artista con id: ", artistaId);
 		Optional<ArtistaEntity> artistaEntity = artistaRepository.findById(artistaId);
-		
-		//TODO
 		if (artistaEntity.isEmpty())
 			throw new EntityNotFoundException("ARTISTA NOT FOUND");
 
+		if (artista.getFechaNacimiento() == null)
+			throw new IllegalOperationException("Fecha Nacimiento is not valid");
+		
+		if (!validateFechaNacimiento(artista.getFechaNacimiento()))
+			throw new IllegalOperationException("Fecha Nacimiento is not valid");
+		
+		if (artista.getFechaFallecimiento() == null)
+			throw new IllegalOperationException("Fecha Fallecimiento is not valid");
+		
+		if (!validateFechaFallecimiento(artista.getFechaFallecimiento()))
+			throw new IllegalOperationException("Fecha Fallecimiento is not valid");	
+		
+		if (artista.getLugarNacimiento() == null)
+			throw new IllegalOperationException("Lugar Nacimiento is not valid");	
+		
+		if (artista.getLugarFallecimiento() == null)
+			throw new IllegalOperationException("Lugar Fallecimiento is not valid");
+		
+		if (artista.getMovimientos() == null)
+			throw new IllegalOperationException("Movimientos is not valid");
+		
+		if (artista.getMuseos() == null)
+			throw new IllegalOperationException("Museos is not valid");
+		
+		if (artista.getObras() == null)
+			throw new IllegalOperationException("Obras is not valid");
+		
+
 		artista.setId(artistaId);
-		log.info("Termina proceso de actualizar el libro con id: " + artistaId);
+		log.info("Termina proceso de actualizar el artista con id: ", artistaId);
 		return artistaRepository.save(artista);
 	}
 	
@@ -104,8 +164,42 @@ public class ArtistaService {
 		if (artistaEntity.isEmpty())
 			throw new EntityNotFoundException("ARTISTA NOT FOUND");
 		
+		List<ObraEntity> obras = artistaEntity.get().getObras();
+		if (!obras.isEmpty())
+			throw new IllegalOperationException("Unable to delete arista because it has associated obras");
+		
 		artistaRepository.deleteById(artistaId);
 		log.info("Termina proceso de borrar el libro con id: " + artistaId);
 	}
 
+	
+	/**
+	 * Verifica que el Nombre sea valido.
+	 *
+	 * @param nombre que se debe verificar
+	 * @return true si el ISBN es valido.
+	 */
+	private boolean validateNombre(String nombre) {
+		return !(nombre == null || nombre.isEmpty());
+	}
+	
+	/**
+	 * Verifica que la Fecha de Nacimiento sea valida.
+	 *
+	 * @param fecha que se debe verificar
+	 * @return true si la Fecha de Nacimiento es valida.
+	 */
+	private boolean validateFechaNacimiento(Date fecha) {
+		return true;
+	}
+	
+	/**
+	 * Verifica que la Fecha de Fallecimiento sea valida.
+	 *
+	 * @param fecha que se debe verificar
+	 * @return true si la Fecha de Fallecimiento es valida.
+	 */
+	private boolean validateFechaFallecimiento(Date fecha) {
+		return true;
+	}
 }
